# api/openapi_v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cineca Agentic Platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/jobs/{job_id}": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Get job status",
        "description": "Retrieve the current status and result of a background job by its identifier. Requires an authenticated user with the `admin` scope. Returns the job `id`, a short `status` (for example `queued`, `running`, `finished`, `failed`, or `cancelled`), and an optional `result` payload when available. Use this endpoint to poll job progress after creating a job.",
        "operationId": "get_job_v1_jobs__job_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": [
          "Jobs"
        ],
        "summary": "Cancel job (idempotent)",
        "description": "Request cancellation of a background job. Requires `admin` scope. This endpoint performs a best-effort cancel: if the job is `queued` or `running` it will be marked `cancelled`. Returns 204 No Content.",
        "operationId": "cancel_job_v1_jobs__job_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/v1/jobs": {
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Create a background job (canonical)",
        "description": "Create a background job request. Accepts a `JobRequest` with a `type` and an optional `payload`. Returns HTTP 202 Accepted with a `Location` header pointing to `/v1/jobs/{job_id}` where callers can poll status. This endpoint supports an `Idempotency-Key` header: if the same key is used, the server will return the same job id (24h dedup). Requires `admin` scope.",
        "operationId": "create_job_v1_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/v1/admin/jobs": {
      "get": {
        "tags": [
          "Admin • Jobs"
        ],
        "summary": "List jobs",
        "description": "Return a paginated list of background jobs managed by the platform.\n\nAdmin-facing endpoint that returns job identifiers and their current status. The\nresponse is paginated and follows a simple `items` / `next_page_token` shape. This proxy\nwraps the demo in-memory job store for local testing; real deployments will map to the\nplatform's persistent job service.",
        "operationId": "list_jobs_v1_admin_jobs_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          },
          {
            "name": "page_token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Page Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "x-internal": true
      },
      "post": {
        "tags": [
          "Admin • Jobs"
        ],
        "summary": "Create a background job (admin proxy)",
        "description": "Create a background job via the canonical job creation endpoint.\n\n    Admin proxy that forwards the incoming request body to the canonical job creation\n    implementation. This route mirrors `/v1/jobs:create` but is exposed under the admin\n    mount so administrative tools can create jobs with the same semantics and authentication\n    as normal clients. The response and headers are passed through from the canonical\n    handler.",
        "operationId": "create_job_proxy_v1_admin_jobs_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        },
        "x-internal": true
      }
    },
    "/v1/admin/jobs/{job_id}": {
      "delete": {
        "tags": [
          "Admin • Jobs"
        ],
        "summary": "Cancel job (admin proxy)",
        "description": "Cancel an existing background job by id (admin proxy).\n\n    Proxy endpoint that delegates cancellation requests to the canonical job cancellation\n    handler. Intended for administrative tooling to cancel jobs on behalf of users or to\n    perform emergency cleanup. Returns the canonical handler's response.",
        "operationId": "cancel_job_proxy_v1_admin_jobs__job_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "x-internal": true
      }
    },
...
    "/v1/internal/db/jobs": {
      "post": {
        "tags": [
          "Internal • DB"
        ],
        "summary": "Create DB job (create|populate)",
        "operationId": "create_db_job_v1_internal_db_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DBJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "security": [
          {
            "oauth2": [
              "admin"
            ]
          }
        ],
        "x-internal": true
      }
    },
    "/v1/internal/db/jobs/{job_id}": {
      "get": {
        "tags": [
          "Internal • DB"
        ],
        "summary": "Get DB job status",
        "description": "Retrieve DB job status by id.",
        "operationId": "get_db_job_v1_internal_db_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "security": [
          {
            "oauth2": [
              "admin"
            ]
          }
        ],
        "x-internal": true
      },
      "delete": {
        "tags": [
          "Internal • DB"
        ],
        "summary": "Cancel DB job",
        "description": "Cancel a DB job; idempotent, returns 204.",
        "operationId": "delete_db_job_v1_internal_db_jobs__job_id__delete",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "security": [
          {
            "oauth2": [
              "admin"
            ]
          }
        ],
        "x-internal": true
      }
    },
    "/v1/internal/db/counts": {
      "get": {
        "tags": [
          "Internal • DB"
        ],
        "summary": "Db Counts",
        "operationId": "db_counts_v1_internal_db_counts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "security": [
          {
            "oauth2": [
              "admin"
            ]
          }
        ],
        "x-internal": true
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DBJobRequest": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "description": "One of: create, populate"
          },
          "wipe": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wipe",
            "description": "When type=create, whether to wipe existing DB"
          },
          "users": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Users",
            "description": "When type=populate, number of synthetic users to generate"
          }
        },
        "type": "object",
        "required": [
          "type"
        ],
        "title": "DBJobRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Job": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status"
        ],
        "title": "Job"
      },
      "JobList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Job"
            },
            "type": "array",
            "title": "Items"
          },
          "next_page_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Page Token"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "JobList"
      },
      "JobRequest": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Logical job type identifier (e.g., 'db.populate')"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload",
            "description": "Arbitrary JSON payload passed to the background worker"
          }
        },
        "type": "object",
        "required": [
          "type"
        ],
        "title": "JobRequest"
      },
      "JobResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Job identifier"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Short job lifecycle status (queued, running, finished, failed, cancelled)"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result",
            "description": "Optional result payload produced by the job when complete"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status"
        ],
        "title": "JobResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "traceId": {
            "type": "string",
            "nullable": true
          },
          "extensions": {
            "type": "object",
            "additionalProperties": true,
            "nullable": true
          }
        },
        "required": [
          "status"
        ],
        "example": {
          "type": "about:blank",
          "title": "Not Found",
          "status": 404,
          "detail": "Item not found",
          "traceId": "..."
        }
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/v1/auth/token"
          }
        }
      },
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "clientCredentials": {
            "tokenUrl": "/v1/auth/token",
            "scopes": {
              "admin": "Admin operations",
              "internal": "Internal-only operations",
              "user": "User-level actions"
            }
          },
          "authorizationCode": {
            "authorizationUrl": "/v1/auth/authorize",
            "tokenUrl": "/v1/auth/token",
            "scopes": {
              "openid": "OpenID Connect scope",
              "profile": "Profile info",
              "email": "Email info"
            }
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad Request",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        }
      },
      "NotFound": {
        "description": "Not Found",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        }
      },
...
      "InternalError": {
        "description": "Internal Server Error",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "/{version}",
      "variables": {
        "version": {
          "default": "v1",
          "enum": [
            "v1",
            "v2"
          ]
        }
      }
    }
  ],
...
}